Rename Acme example to Proprietary and modify package to add the prefix com.example.
