--- acme-javaee-servlet/pom.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <properties>
-        <version.windup>2.3.0-SNAPSHOT</version.windup>
-    </properties>
-
-    <groupId>org.jboss.windup.quickstarts</groupId>
-    <artifactId>windup-acme-javaee-servlet</artifactId>
-    <version>2.3.0-SNAPSHOT</version>
-    <packaging>pom</packaging>
-
-    <name>Windup Quickstart: Acme to Java EE Servlet</name>
-
-    <modules>
-        <module>rules-java</module>
-    </modules>
-
-    <licenses>
-        <license>
-            <name>Apache License, Version 2.0</name>
-            <distribution>repo</distribution>
-            <url>http://www.apache.org/licenses/LICENSE-2.0.html</url>
-        </license>
-    </licenses>
-
-    <scm>
-        <tag>master</tag>
-        <connection>scm:git:https://github.com/windup/windup-quickstarts.git</connection>
-        <developerConnection>scm:git:git@github.com:windup/windup-quickstarts.git</developerConnection>
-        <url>http://github.com/windup/windup-quickstarts</url>
-    </scm>
-
-</project>
