+++ proprietary-javaee-servlet/README.adoc
@@ -1,0 +1,29 @@
+[[proprietary-javaee-servlet-windup-rule-that-detects-proprietary-proprietary-servlet-annotations]]
+= proprietary-javaee-servlet: Windup Rule That Detects Proprietary Proprietary Servlet Annotations
+
+Author: Ondra Zizka +
+Level: Beginner +
+Summary: Windup rule that reports on Proprietary servlet annotations +
+Source: https://github.com/windup/windup-quickstarts/ +
+
+[[what-is-it]]
+== What is it?
+
+In this example, the fictional "Proprietary" corporation provides its own proprietary servlet and filter annotations for dependency injection. 
+If your application uses them, they must be replaced with the standard Java EE 6 annotations. 
+
+These examples demonstrates how to create a Windup rule add-on that searches for these proprietary annotations and reports on them.
+
+The rule searches for the following annotations:
+
+* _@ProprietaryServlet_: This is the equivalent of the Java EE 6 _@WebServlet_ annotation.
+* _@ProprietaryFilter_: This is the equivalent of the Java EE 6 _@WebFilter_ annotation.
+* _@ProprietaryInitParam_: This is the equivalent of the Java EE 6 _@WebInitParam_ annotation.
+
+You can create Windup rule add-ons using Java or XML. This quickstart provides examples of both types of rules.
+
+* link:rules-java/README.adoc[Java rule add-on]: Follow the instructions here to see how to create a Java rule add-on.
+* link:rules-xml/README.adoc[XML rule]: Follow the instructions here to see how to create an XML rule.
+
+NOTE: If you deploy both the Java-based rule add-on and XML rule at the same time, you will see double the story points!
+
