+++ proprietary-javaee-servlet/pom.xml
@@ -7,24 +7,15 @@
         <version.windup>2.3.0-SNAPSHOT</version.windup>
     </properties>
 
-    <parent>
-        <groupId>org.jboss</groupId>
-        <artifactId>jboss-parent</artifactId>
-        <version>16</version>
-        <relativePath />
-    </parent>
-
     <groupId>org.jboss.windup.quickstarts</groupId>
-    <artifactId>windup-quickstarts-root</artifactId>
+    <artifactId>windup-proprietary-javaee-servlet</artifactId>
     <version>2.3.0-SNAPSHOT</version>
     <packaging>pom</packaging>
 
-    <name>Windup QuickStart Root</name>
+    <name>Windup Quickstart: Proprietary to Java EE Servlet</name>
 
     <modules>
-        <module>proprietary-javaee-servlet</module>
-        <module>ejb-beanutils-async</module>
-        <module>victi.ms</module>
+        <module>rules-java</module>
     </modules>
 
     <licenses>
