+++ proprietary-javaee-servlet/rules-java/src/main/resources/META-INF/beans.xml
